requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.4.3-cp37-cp37m-win_amd64.whl
pytest-allure-adaptor